Move lib to new nx tsc lib
