+++ core-web/libs/dotcms/project.json
@@ -1,27 +1,36 @@
 {
     "$schema": "../../node_modules/nx/schemas/project-schema.json",
-    "projectType": "library",
     "sourceRoot": "libs/dotcms/src",
-    "prefix": "dotcms",
+    "projectType": "library",
     "targets": {
         "build": {
-            "executor": "@nrwl/angular:ng-packagr-lite",
+            "executor": "@nrwl/js:tsc",
+            "outputs": ["{options.outputPath}"],
             "options": {
+                "outputPath": "dist/libs/dotcms",
+                "main": "libs/dotcms/src/index.ts",
                 "tsConfig": "libs/dotcms/tsconfig.lib.json",
-                "project": "libs/dotcms/ng-package.json"
-            },
-            "configurations": {
-                "production": {
-                    "tsConfig": "libs/dotcms/tsconfig.lib.prod.json"
-                }
+                "assets": ["libs/dotcms/*.md"]
             }
+        },
+        "publish": {
+            "executor": "@nrwl/workspace:run-commands",
+            "options": {
+                "command": "node tools/scripts/publish.mjs dotcms {args.ver} {args.tag}"
+            },
+            "dependsOn": [
+                {
+                    "projects": "self",
+                    "target": "build"
+                }
+            ]
         },
         "lint": {
             "executor": "@nrwl/linter:eslint",
+            "outputs": ["{options.outputFile}"],
             "options": {
-                "lintFilePatterns": ["libs/dotcms/src/**/*.ts", "libs/dotcms/src/**/*.html"]
-            },
-            "outputs": ["{options.outputFile}"]
+                "lintFilePatterns": ["libs/dotcms/**/*.ts"]
+            }
         },
         "test": {
             "executor": "@nrwl/jest:jest",
