+++ core-web/libs/dotcms/src/index.ts
@@ -10,19 +10,15 @@ import { DotApiWidget } from './lib/api/DotApiWidget';
 import { DotCMSHttpClient } from './lib/utils/DotCMSHttpClient';
 import { DotCMSConfigurationParams, DotCMSFormConfig } from './lib/models';
 import { DotApiContent } from './lib/api/DotApiContent';
-import { DotApiContentType } from './lib/api/DotApiContentType';
-import { DotApiForm } from './lib/api/DotApiForm';
 
 export interface DotCMSApp {
     auth: DotApiAuthorization;
     content: DotApiContent;
-    contentType: DotApiContentType;
     esSearch: DotApiElasticSearch;
     event: DotApiEvent;
     nav: DotApiNavigation;
     page: DotApiPage;
     site: DotApiSite;
-    form: (formConfig: DotCMSFormConfig, win?: Window) => DotApiForm;
     widget: DotApiWidget;
     config: DotApiConfiguration;
     language: DotApiLanguage;
@@ -34,16 +30,13 @@ export const initDotCMS = (config: DotCMSConfigurationParams): DotCMSApp => {
     const apiConfig = new DotApiConfiguration(httpClient);
     const apiLanguage = new DotApiLanguage(apiConfig);
     const content = new DotApiContent(httpClient);
-    const dotApiContentType = new DotApiContentType(httpClient);
 
     return {
         auth: new DotApiAuthorization(),
         config: apiConfig,
         content: content,
-        contentType: new DotApiContentType(httpClient),
         esSearch: new DotApiElasticSearch(httpClient),
         event: new DotApiEvent(),
-        form: (formConfig: DotCMSFormConfig) => new DotApiForm(dotApiContentType, formConfig),
         language: apiLanguage,
         nav: new DotApiNavigation(httpClient),
         page: new DotApiPage(httpClient, apiLanguage),
