+++ core-web/libs/dotcms/src/lib/api/DotApiAuthorization.ts
@@ -1,5 +1,4 @@
 import { DotCMSAuthorizationLoginParams, DotCMSError } from '../models';
-import fetch, { Response } from 'cross-fetch';
 
 function getErrorMessage(data: { [key: string]: any }) {
     if (data.errors) {
