+++ core-web/libs/dotcms/src/lib/api/DotApiElasticSearch.ts
@@ -1,4 +1,4 @@
-import { DotCMSError } from './../models/DotCMSError.model';
+import { DotCMSError } from '../models/DotCMSError.model';
 import { DotCMSHttpClient } from '../utils/DotCMSHttpClient';
 import { getEsQuery } from '../utils/getEsQuery';
 import { DotCMSElasticSearchResult, DotCMSElasticSearchParams } from '../models';
