+++ core-web/libs/dotcms/src/lib/utils/request.ts
@@ -1,12 +1,11 @@
 import { DotCMSConfigurationParams, DotAppHttpRequestParams } from '../models';
-import fetch from 'cross-fetch';
 
 async function getUrl(
     { params, url }: DotAppHttpRequestParams,
     { host, hostId }: DotCMSConfigurationParams
 ): Promise<string> {
     if (!host) {
-        throw new Error('Please pass the DotCMS instance in the initDotCMS initialization')
+        throw new Error('Please pass the DotCMS instance in the initDotCMS initialization');
     }
 
     const newUrl = new URL(`${host}${url}`);
@@ -15,7 +14,7 @@ async function getUrl(
     if (paramsKeys.length) {
         paramsKeys.map((key: string) => {
             newUrl.searchParams.append(key, params[key]);
-        })
+        });
     }
 
     if (hostId) {
