+++ core-web/libs/dotcms/tsconfig.json
@@ -1,5 +1,8 @@
 {
     "extends": "../../tsconfig.base.json",
+    "compilerOptions": {
+        "module": "commonjs"
+    },
     "files": [],
     "include": [],
     "references": [
@@ -9,8 +12,5 @@
         {
             "path": "./tsconfig.spec.json"
         }
-    ],
-    "compilerOptions": {
-        "target": "es2020"
-    }
+    ]
 }
