+++ core-web/libs/dotcms/tsconfig.lib.json
@@ -2,21 +2,12 @@
     "extends": "./tsconfig.json",
     "compilerOptions": {
         "outDir": "../../dist/out-tsc",
-        "target": "es2015",
         "declaration": true,
-        "declarationMap": true,
-        "inlineSources": true,
         "types": [],
-        "lib": ["dom", "es2018"],
         "paths": {
-            "@dotcms/*": ["dist/libs/*"]
+            "dotcms-models": ["libs/dotcms-models/src/index.ts"]
         }
     },
-    "angularCompilerOptions": {
-        "skipTemplateCodegen": true,
-        "strictMetadataEmit": true,
-        "enableResourceInlining": true
-    },
-    "exclude": ["src/test-setup.ts", "**/*.spec.ts", "**/*.test.ts", "jest.config.ts"],
-    "include": ["**/*.ts"]
+    "include": ["**/*.ts"],
+    "exclude": ["jest.config.ts", "**/*.spec.ts", "**/*.test.ts"]
 }
