+++ core-web/libs/dotcms/tsconfig.spec.json
@@ -5,6 +5,5 @@
         "module": "commonjs",
         "types": ["jest", "node"]
     },
-    "files": ["src/test-setup.ts"],
-    "include": ["**/*.spec.ts", "**/*.test.ts", "**/*.d.ts", "jest.config.ts"]
+    "include": ["jest.config.ts", "**/*.test.ts", "**/*.spec.ts", "**/*.d.ts"]
 }
