--- core-web/libs/dotcms/typedoc.json
@@ -1,10 +1,0 @@
-{
-    "name": "DotCMS",
-    "out": "../../dist-docs/docs/dotcms",
-    "entryPoints": ["./src/index.ts"],
-    "readme": "./libs/dotcms/README.md",
-    "excludePrivate": true,
-    "tsconfig": "./tsconfig.lib.json",
-    "excludeExternals": true,
-    "exclude": ["**/*.spec.ts"]
-}
