+++ core-web/tsconfig.base.json
@@ -23,7 +23,6 @@
             "@dotcms/block-editor/services": ["libs/block-editor/src/lib/shared/services/index.ts"],
             "@dotcms/dot-layout-grid": ["libs/dot-layout-grid/src/public_api.ts"],
             "@dotcms/dot-rules": ["libs/dot-rules/src/public_api.ts"],
-            "@dotcms/dotcms": ["libs/dotcms/src/index.ts"],
             "@dotcms/dotcms-field-elements": ["libs/dotcms-field-elements/src/index.ts"],
             "@dotcms/dotcms-field-elements/loader": ["dist/libs/dotcms-field-elements/loader"],
             "@dotcms/dotcms-js": ["libs/dotcms-js/src/public_api.ts"],
@@ -37,7 +36,8 @@
             "@portlets/*": ["apps/dotcms-ui/src/app/portlets/*"],
             "@services/*": ["apps/dotcms-ui/src/app/api/services/*"],
             "@shared/*": ["apps/dotcms-ui/src/app/shared/*"],
-            "@tests/*": ["apps/dotcms-ui/src/app/test/*"]
+            "@tests/*": ["apps/dotcms-ui/src/app/test/*"],
+            "dotcms": ["libs/dotcms/src/index.ts"]
         }
     },
     "exclude": ["node_modules", "tmp"]
